backend/pyproject.toml
[project]
name = "flash-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "firebase-admin>=7.0.0",
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "google-genai>=1.27.0",
    "pyrebase4>=4.8.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "setuptools>=75.0.0",
]
